We don't use TermVectors API anymore

+++ core/src/main/java/br/ufmg/dcc/latin/searcher/matching/DFIScorer.java
@@ -22,7 +22,7 @@ public class DFIScorer extends DFISimilarity implements Scorer{
 	}
 
 	@Override
-	public float totalScore(BasicStats[] basicStats, long[] termFreq, long docLen, int termCount) {
+	public float totalScore(BasicStats[] basicStats, int[] termFreq, long docLen, int termCount) {
 		float s = 0;
 		for (int i = 0; i < termCount; i++) {
 			// we need to encode a normalization as lucene does that too
