+++ core/src/main/java/br/ufmg/dcc/latin/searcher/matching/LMDirichletScorer.java
@@ -19,7 +19,7 @@ public class LMDirichletScorer extends LMDirichletSimilarity implements Scorer {
 	}
 
 	@Override
-	public float totalScore(BasicStats[] basicStats, long[] termFreq, long docLen, int termCount) {
+	public float totalScore(BasicStats[] basicStats, int[] termFreq, long docLen, int termCount) {
 		float s = 0;
 		for (int i = 0; i < termCount; i++) {
 			LMStats lmStats = new LMStats("lm");
